ETDA-1518 Refactored some tests (admins tests in the author tests for some reason).  Added a line space in appilication config
