+++ config/application.rb
@@ -29,9 +29,6 @@ module EtdaWorkflow
     # Initialize configuration defaults for originally generated Rails version.
 
     config.load_defaults "6.0"
-    # Rails 6 uses zeitwerk's autoloader as default
-    # Use classic autoloader to avoid errors
-    # config.autoloader = :classic
 
     # Logging
     logging_config = Rails.application.config_for(:logging)
@@ -67,6 +64,7 @@ module EtdaWorkflow
     config.autoload_paths += [
       Rails.root.join('app/presenters')
     ]
+
     config.autoload_paths += Dir["#{config.root}/lib"]
 
     config.autoload_paths += Dir[File.join(Rails.root, "lib", "core_ext", "*.rb")].each {|l| require l }
